EM-2919: Move icp interfaces and remove screen from icp store
