+++ projects/media-viewer/src/lib/store/reducers/reducers.ts
@@ -1,5 +1,4 @@
-import { createFeatureSelector, ActionReducerMap } from '@ngrx/store';
-
+import { ActionReducerMap, createFeatureSelector } from '@ngrx/store';
 import * as fromDocument from './document.reducer';
 import * as fromAnnotation from './annotatons.reducer';
 import * as fromTags from './tags.reducer';
@@ -7,6 +6,7 @@ import * as fromBookmarks from './bookmarks.reducer';
 import { BookmarksState } from '../../viewers/pdf-viewer/side-bar/bookmarks/bookmarks.interfaces';
 import * as fromRedaction from './redaction.reducer';
 import * as fromIcp from './icp.reducer';
+import { IcpState } from '../../icp/icp.interfaces';
 
 export interface State {
   document: fromDocument.DocumentState;
@@ -14,7 +14,7 @@ export interface State {
   tags: fromTags.TagsState;
   bookmarks: BookmarksState;
   redactions: fromRedaction.RedactionState;
-  icp: fromIcp.IcpState;
+  icp: IcpState;
 }
 
 export const reducers: ActionReducerMap<State> = {
